chore: update create-vite templates version
