+++ packages/plugin-legacy/package.json
@@ -10,8 +10,7 @@
   "main": "index.js",
   "types": "index.d.ts",
   "scripts": {
-    "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s --commit-path . --lerna-package plugin-legacy",
-    "release": "ts-node ../../scripts/release.ts"
+    "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s --commit-path . --lerna-package plugin-legacy"
   },
   "engines": {
     "node": ">=12.0.0"
